pyproject.toml
[project]
name = "graphrapping"
version = "0.1.0"
description = "Beauty product semantic signal graph: review NER/BEE/REL → canonical facts → serving signals → recommendation"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
